Moved elevator code to subsystems folder
